requirements.txt
tweepy>=4.0
unidecode
Mastodon.py>=1.7.0
bs4